Trim eager resource and panel imports from app startup
