+++ apps/app/src/main.tsx
@@ -2,7 +2,6 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { BrowserRouter } from "react-router-dom";
-import { disableGlobalCursorStyles } from "react-resizable-panels";
 import { App } from "./App";
 import { AppErrorBoundary } from "./components/AppErrorBoundary";
 import { AppToaster } from "./components/AppToaster";
@@ -25,7 +24,6 @@ registerProviderCliInstallQueryClient(appQueryClient);
 initializePreferredTheme();
 applyCachedAppThemeCss();
 initializeFavicon();
-disableGlobalCursorStyles();
 
 createRoot(document.getElementById("root")!, {
   onUncaughtError: (error, errorInfo) => {
